Provide direct access to the token response
The OIDC protocol allows specifying alternative response_type that can
include the id_token in addition to the access token. For clients
utilizing this they can avoid an additional request to the userinfo
endpoint to access id_token information so it can be advantageous to
provide direct access to the response json for clients wanting to make
use of this performance improvement.